requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
temporalio>=1.5.0
redis>=5.0.0
opencv-python-headless>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
pyyaml>=6.0
loguru>=0.7.0
boto3>=1.28.0
pillow>=10.0.0
